change project directory to multi project
