--- .scalafmt.conf
@@ -1,8 +1,0 @@
-# 詳細な設定はhttps://scalameta.org/scalafmt/
-# 使用する方言（Scala 2.13 の場合）
-runner.dialect = scala213
-
-# 最大行長
-maxColumn = 120
-# メソッド定義の改行
-verticalMultiline.atDefnSite = true
