--- app/controllers/HomeController.scala
@@ -1,13 +1,0 @@
-package controllers
-
-import play.api.mvc._
-
-import javax.inject._
-
-@Singleton
-class HomeController @Inject()(val controllerComponents: ControllerComponents) extends BaseController {
-
-    def index(): Action[AnyContent] = Action { implicit request: Request[AnyContent] =>
-        Ok(views.html.index())
-    }
-}
