--- app/sample/SampleCaseClass.scala
@@ -1,3 +1,0 @@
-package sample
-
-case class SampleCaseClass(arg1: String, arg2: String)
