--- app/views/index.scala.html
@@ -1,3 +1,0 @@
-@main("Welcome to Play") {
-  <h1>Welcome to Play!</h1>
-}
