--- app/views/main.scala.html
@@ -1,19 +1,0 @@
-@(title: String)(content: Html)
-
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        @* Here's where we render the page title `String`. *@
-        <title>@title</title>
-        <link rel="stylesheet" media="screen" href="@routes.Assets.versioned("stylesheets/main.css")">
-        <link rel="shortcut icon" type="image/png" href="@routes.Assets.versioned("images/favicon.png")">
-
-    </head>
-    <body>
-        @* And here's where we render the `Html` object containing
-         * the page content. *@
-        @content
-
-      <script src="@routes.Assets.versioned("javascripts/main.js")" type="text/javascript"></script>
-    </body>
-</html>
