+++ project/plugins.sbt
@@ -1,7 +1,5 @@
 addSbtPlugin("org.playframework" % "sbt-plugin" % "3.0.5")
 addSbtPlugin("org.foundweekends.giter8" % "sbt-giter8-scaffold" % "0.16.2")
-// todo: コードフォーマットプラグインを追加
-addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.4.2")
 // Linter
 addSbtPlugin("ch.epfl.scala" % "sbt-scalafix" % "0.10.4")
 // カバレッジ
