+++ study-scala-app/app/model/entity/User.scala
