+++ study-scala-app/app/model/table/UserTable.scala
