+++ study-scala-app/app/repository/UserRepository.scala
