+++ study-scala-app/app/service/UserService.scala
