+++ study-scala-common/.bsp/sbt.json
@@ -1,0 +1,1 @@
+{"name":"sbt","version":"1.10.2","bspVersion":"2.1.0-M1","languages":["scala"],"argv":["C:\\Program Files\\Java\\jdk-19/bin/java","-Xms100m","-Xmx100m","-classpath","C:/Users/nakai/AppData/Roaming/JetBrains/IdeaIC2024.1/plugins/Scala/launcher/sbt-launch.jar","-Dsbt.script=C:\\tool\\sbt\\bin\\sbt.bat","xsbt.boot.Boot","-bsp"]}
