+++ study-scala-common/app/sample/SampleAbstractClass.scala
