+++ study-scala-common/app/sample/SampleCaseClass.scala
@@ -1,0 +1,3 @@
+package sample
+
+case class SampleCaseClass(arg1: String, arg2: String)
