+++ study-scala-common/app/sample/SampleClass.scala
