+++ study-scala-common/app/sample/SampleFunctionClass.scala
