+++ study-scala-common/app/sample/SampleFuture.scala
