+++ study-scala-common/app/sample/SampleTrait.scala
