+++ study-scala-common/app/sample/SpecialObject.scala
