+++ study-scala-common/app/sample/UnarySample.scala
