+++ study-scala-common/conf/messages
@@ -1,0 +1,1 @@
+# https://www.playframework.com/documentation/latest/ScalaI18N
